oursource fen/state utility functions
